# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: 微服务
  version: ""
host: localhost:8000
basePath: /api
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /actual_time/get_actual_time_data:
    get:
      tags:
      - ActualTime
      summary: 接收数据库监听得到的数据
      description: 接收数据库监听得到的数据
      operationId: ActualTime#GetActualTimeData
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ActualTimeGetActualTimeDataResponseBody'
            required:
            - errcode
            - errmsg
            - data
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ActualTimeGetActualTimeDataBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ActualTimeGetActualTimeDataInternalServerErrorResponseBody'
      schemes:
      - http
  /actual_time/receive_third_party_push_data:
    post:
      tags:
      - ActualTime
      summary: 接收第三方推送数据
      description: 接收第三方推送数据--大厅排队办事实时图基础数据
      operationId: ActualTime#ReceiveThirdPartyPushData
      parameters:
      - name: ReceiveThirdPartyPushDataRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/ActualTimeReceiveThirdPartyPushDataRequestBody'
          required:
          - methodName
          - count
          - data
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ActualTimeReceiveThirdPartyPushDataResponseBody'
            required:
            - errcode
            - errmsg
            - result
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ActualTimeReceiveThirdPartyPushDataBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ActualTimeReceiveThirdPartyPushDataInternalServerErrorResponseBody'
      schemes:
      - http
...
  /third_part/get_hall_management_data:
    get:
      tags:
      - thirdPart
      summary: 接收大厅管理的数据
      description: 接收大厅管理的数据
      operationId: thirdPart#GetActualTimeData
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ThirdPartGetActualTimeDataResponseBody'
            required:
            - errcode
            - errmsg
            - data
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ThirdPartGetActualTimeDataBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ThirdPartGetActualTimeDataInternalServerErrorResponseBody'
      schemes:
      - http
  /third_part/gorm_related_search:
    get:
      tags:
      - thirdPart
      summary: gorm关联查询
      description: gorm关联查询
      operationId: thirdPart#GormRelatedSearch
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ThirdPartGormRelatedSearchResponseBody'
            required:
            - errcode
            - errmsg
            - data
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ThirdPartGormRelatedSearchBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ThirdPartGormRelatedSearchInternalServerErrorResponseBody'
      schemes:
      - http
...
definitions:
  ActualTimeGetActualTimeDataBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GetActualTimeData_bad_request_Response_Body result type (default
      view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ActualTimeGetActualTimeDataInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GetActualTimeData_internal_server_error_Response_Body result type
      (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ActualTimeGetActualTimeDataResponseBody:
    title: ActualTimeGetActualTimeDataResponseBody
    type: object
    properties:
      data:
        $ref: '#/definitions/CanalDataRespResponseBody'
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    example:
      data:
        dataInfo: 1035
        dataType: 1035
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
    - data
  ActualTimeReceiveThirdPartyPushDataBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: ReceiveThirdPartyPushData_bad_request_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ActualTimeReceiveThirdPartyPushDataInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: ReceiveThirdPartyPushData_internal_server_error_Response_Body result
      type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ActualTimeReceiveThirdPartyPushDataRequestBody:
    title: ActualTimeReceiveThirdPartyPushDataRequestBody
    type: object
    properties:
      count:
        type: integer
        description: 数据数量
        example: 21
        format: int64
      data:
        description: 第三方推送数据
        example: ""
      methodName:
        type: integer
        description: 推送的具体方法
        example: 1
        format: int64
    example:
      count: 21
      data: ""
      methodName: 1
    required:
    - methodName
    - count
    - data
  ActualTimeReceiveThirdPartyPushDataResponseBody:
    title: ActualTimeReceiveThirdPartyPushDataResponseBody
    type: object
    properties:
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
      result:
        type: string
        description: success
        example: getDataSuccess
    example:
      errcode: 0
      errmsg: ""
      result: getDataSuccess
    required:
    - errcode
    - errmsg
    - result
  CanalDataRespResponseBody:
    title: 'Mediatype identifier: canaldataresp; view=default'
    type: object
    properties:
      dataInfo:
        description: 数据
        example: 1035
      dataType:
        type: integer
        description: 数据类别
        example: 1035
        format: int32
    description: 数据库增量数据 (default view)
    example:
      dataInfo: 1035
      dataType: 1035
    required:
    - dataType
    - dataInfo
...
  CompanyProfileRespResponseBody:
    title: 'Mediatype identifier: companyprofileresp; view=default'
    type: object
    properties:
      industry:
        type: integer
        description: 行业Id
        example: 1036
        format: int64
      name:
        type: string
        description: 公司名字
        example: 青朵教育
      userId:
        type: string
        description: 法人名字
        example: 王大锤
    description: 公司简洁
    example:
      industry: 1036
      name: 青朵教育
      userId: 王大锤
    required:
    - industry
    - name
    - userId
...
  HallManagementRespResponseBody:
    title: 'Mediatype identifier: hallmanagementresp; view=default'
    type: object
    properties:
      cardNum:
        type: string
        description: 身份证
        example: "1035"
      name:
        type: string
        description: 名字
        example: 小白
      ouName:
        type: string
        description: 部门名字
        example: 社保
    description: 大厅管理系统数据 (default view)
    example:
      cardNum: "1035"
      name: 小白
      ouName: 社保
    required:
    - cardNum
    - name
    - ouName
  LegalPersonUserRespResponseBody:
    title: 'Mediatype identifier: legalpersonuserresp; view=default'
    type: object
    properties:
      companies:
        type: array
        items:
          $ref: '#/definitions/CompanyProfileRespResponseBody'
        description: 公司
        example:
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
      id:
        type: integer
        description: 用户Id
        example: 1035
        format: int64
      name:
        type: string
        description: 名字
        example: 小白
    description: 法人用户列表 (default view)
    example:
      companies:
      - industry: 1036
        name: 青朵教育
        userId: 王大锤
      - industry: 1036
        name: 青朵教育
        userId: 王大锤
      id: 1035
      name: 小白
    required:
    - id
    - name
    - companies
...
  ThirdPartGetActualTimeDataBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: GetActualTimeData_bad_request_Response_Body result type (default
      view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ThirdPartGetActualTimeDataInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GetActualTimeData_internal_server_error_Response_Body result type
      (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ThirdPartGetActualTimeDataResponseBody:
    title: ThirdPartGetActualTimeDataResponseBody
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/HallManagementRespResponseBody'
        example:
        - cardNum: "1035"
          name: 小白
          ouName: 社保
        - cardNum: "1035"
          name: 小白
          ouName: 社保
        - cardNum: "1035"
          name: 小白
          ouName: 社保
        - cardNum: "1035"
          name: 小白
          ouName: 社保
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    example:
      data:
      - cardNum: "1035"
        name: 小白
        ouName: 社保
      - cardNum: "1035"
        name: 小白
        ouName: 社保
      - cardNum: "1035"
        name: 小白
        ouName: 社保
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
    - data
  ThirdPartGormRelatedSearchBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: GormRelatedSearch_bad_request_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ThirdPartGormRelatedSearchInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GormRelatedSearch_internal_server_error_Response_Body result type
      (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ThirdPartGormRelatedSearchResponseBody:
    title: ThirdPartGormRelatedSearchResponseBody
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/LegalPersonUserRespResponseBody'
        example:
        - companies:
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          id: 1035
          name: 小白
        - companies:
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          id: 1035
          name: 小白
        - companies:
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          - industry: 1036
            name: 青朵教育
            userId: 王大锤
          id: 1035
          name: 小白
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    example:
      data:
      - companies:
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        id: 1035
        name: 小白
      - companies:
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        id: 1035
        name: 小白
      - companies:
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        id: 1035
        name: 小白
      - companies:
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        - industry: 1036
          name: 青朵教育
          userId: 王大锤
        id: 1035
        name: 小白
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
    - data
...
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: |-
      JWT 认证

      **Security Scopes**:
        * `role:user`: 用户
        * `role:admin`: 管理员
    name: Authorization
    in: header