# gen/http/third_part/server/encode_decode.go
// Code generated by goa v3.2.4, DO NOT EDIT.
//
// thirdPart HTTP server encoders and decoders
//
// Command:
// $ goa gen boot/design

package server

import (
	thirdpart "boot/gen/third_part"
	"context"
	"net/http"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetActualTimeDataResponse returns an encoder for responses returned by
// the thirdPart GetActualTimeData endpoint.
func EncodeGetActualTimeDataResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(*thirdpart.GetActualTimeDataResult)
		enc := encoder(ctx, w)
		body := NewGetActualTimeDataResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeGetActualTimeDataError returns an encoder for errors returned by the
// GetActualTimeData thirdPart endpoint.
func EncodeGetActualTimeDataError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		en, ok := v.(ErrorNamer)
		if !ok {
			return encodeError(ctx, w, v)
		}
		switch en.ErrorName() {
		case "bad_request":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewGetActualTimeDataBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", "bad_request")
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "internal_server_error":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewGetActualTimeDataInternalServerErrorResponseBody(res)
			}
			w.Header().Set("goa-error", "internal_server_error")
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeGormRelatedSearchResponse returns an encoder for responses returned by
// the thirdPart GormRelatedSearch endpoint.
func EncodeGormRelatedSearchResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(*thirdpart.GormRelatedSearchResult)
		enc := encoder(ctx, w)
		body := NewGormRelatedSearchResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeGormRelatedSearchError returns an encoder for errors returned by the
// GormRelatedSearch thirdPart endpoint.
func EncodeGormRelatedSearchError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		en, ok := v.(ErrorNamer)
		if !ok {
			return encodeError(ctx, w, v)
		}
		switch en.ErrorName() {
		case "bad_request":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewGormRelatedSearchBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", "bad_request")
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "internal_server_error":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewGormRelatedSearchInternalServerErrorResponseBody(res)
			}
			w.Header().Set("goa-error", "internal_server_error")
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalThirdpartHallManagementRespToHallManagementRespResponseBody builds a
// value of type *HallManagementRespResponseBody from a value of type
// *thirdpart.HallManagementResp.
func marshalThirdpartHallManagementRespToHallManagementRespResponseBody(v *thirdpart.HallManagementResp) *HallManagementRespResponseBody {
	res := &HallManagementRespResponseBody{
		CardNum: v.CardNum,
		Name:    v.Name,
		OuName:  v.OuName,
	}

	return res
}

// marshalThirdpartLegalPersonUserRespToLegalPersonUserRespResponseBody builds
// a value of type *LegalPersonUserRespResponseBody from a value of type
// *thirdpart.LegalPersonUserResp.
func marshalThirdpartLegalPersonUserRespToLegalPersonUserRespResponseBody(v *thirdpart.LegalPersonUserResp) *LegalPersonUserRespResponseBody {
	res := &LegalPersonUserRespResponseBody{
		ID:   v.ID,
		Name: v.Name,
	}
	if v.Companies != nil {
		res.Companies = make([]*CompanyProfileRespResponseBody, len(v.Companies))
		for i, val := range v.Companies {
			res.Companies[i] = marshalThirdpartCompanyProfileRespToCompanyProfileRespResponseBody(val)
		}
	}

	return res
}

// marshalThirdpartCompanyProfileRespToCompanyProfileRespResponseBody builds a
// value of type *CompanyProfileRespResponseBody from a value of type
// *thirdpart.CompanyProfileResp.
func marshalThirdpartCompanyProfileRespToCompanyProfileRespResponseBody(v *thirdpart.CompanyProfileResp) *CompanyProfileRespResponseBody {
	res := &CompanyProfileRespResponseBody{
		Industry: v.Industry,
		Name:     v.Name,
		UserID:   v.UserID,
	}

	return res
}